+++ src/JoinPolicy.cpp
@@ -9,7 +9,7 @@ MandatesJoinPolicy* MandatesJoinPolicy:: clone() const {
 int MandatesJoinPolicy::choose(vector<int> &offers, Simulation &s) {
     Coalition &coalition = s.getCoalition(offers[0]);
     int index = 0;
-    for (int i = 1; i < (int) offers.size(); i++) {
+    for (unsigned int i = 1; i < offers.size(); i++) {
         Coalition &otherCoalition = s.getCoalition(offers[i]);
         if (otherCoalition.getMandates() > coalition.getMandates()) {
             index = i;
@@ -23,5 +23,5 @@ LastOfferJoinPolicy* LastOfferJoinPolicy:: clone() const {
 }
 
 int LastOfferJoinPolicy::choose(vector<int> &offers, Simulation &s) {
-    return (int) offers.size() - 1;
+    return offers.size() - 1;
 }
